refactor: SpendTx, AENS transaction constructors now do a lot of hard work previously done by Context move GetTTL/GetNextNonce to package transactions
